Delete static_cast

+++ src/population/selector/TournamentSelector.cpp
@@ -17,11 +17,11 @@ Individual TournamentSelector::select(const Individuals& individuals) const {
     return individuals[0];
   }
 
-  unsigned long count = max(individuals.size() / 50, static_cast<unsigned long>(1));
-  
+  unsigned long count = max(individuals.size() / 50, 1UL);
+
   Individuals candidates;
   candidates.reserve(count);
-  
+
   for (unsigned long i = 0; i < count; i++) {
     unsigned long index = numberGenerator->generate() % individuals.size();
     Individual candidate = individuals[index];
